requirements.txt
fastapi
uvicorn[standard]
python-multipart

sqlalchemy
psycopg2-binary


pydantic
python-dotenv